feat: move actions folder to out of app folder
