+++ src/app/actions/products/get-product-action.ts
@@ -1,8 +1,9 @@
 'use server'
 
 import { revalidatePath } from 'next/cache'
+
 import { getProduct } from '@/app/services/api-products'
-import { getErrorMessage } from '@/app/utils/getErrorMessage'
+import { getErrorMessage } from '@/utils/getErrorMessage'
 
 export const getProductServerAction = async (id: string) => {
   try {
