+++ src/app/actions/products/update-product.action.ts
@@ -1,9 +1,10 @@
 'use server'
 
 import { revalidatePath } from 'next/cache'
+
+import { getErrorMessage } from '@/utils/getErrorMessage'
 import { updateProduct } from '@/app/services/api-products'
-import { getErrorMessage } from '@/app/utils/getErrorMessage'
-import { getStringFormDataValue } from '@/app/utils/getStringFormDataValue'
+import { getStringFormDataValue } from '@/utils/getStringFormDataValue'
 
 export const updateProductServerAction = async (formData: FormData) => {
   const id = getStringFormDataValue(formData, '_id')
