+++ src/app/actions/todos/update-todo-action.ts
@@ -3,7 +3,7 @@
 import { revalidatePath } from 'next/cache'
 
 import { updateTodo } from '@/app/services/api-todos'
-import { getErrorMessage } from '@/app/utils/getErrorMessage'
+import { getErrorMessage } from '@/utils/getErrorMessage'
 
 export const updateTodoServerAction = async (id: string) => {
   try {
