+++ src/app/actions/users/get-user-action.ts
@@ -1,8 +1,9 @@
 'use server'
 
-import { getUser } from '@/app/services/api-users'
-import { getErrorMessage } from '@/app/utils/getErrorMessage'
 import { revalidatePath } from 'next/cache'
+
+import { getUser } from '@/app/services/api-users'
+import { getErrorMessage } from '@/utils/getErrorMessage'
 
 export const getUserServerAction = async (id: string) => {
   try {
