+++ src/app/dashboard/products/[id]/form.tsx
@@ -1,14 +1,14 @@
 'use client'
 
 import { useEffect, useState } from 'react'
-import { toast } from 'react-toastify'
 
 import Image from 'next/image'
+import { toast } from 'react-toastify'
 import { useRouter } from 'next/navigation'
 
 import { Product } from '@/types'
+import { convertStringToBoolean } from '@/utils/convertStringToBoolean'
 import SubmitButton from '@/app/ui/dashboard/submit-button/submit-button'
-import { convertStringToBoolean } from '@/app/utils/convertStringToBoolean'
 import { updateProductServerAction } from '@/app/actions/products/update-product.action'
 
 type FormProps = {
