+++ src/app/dashboard/users/usersTableBody.tsx
@@ -6,8 +6,8 @@ import Image from 'next/image'
 
 import { User } from '@/types'
 import DeleteForm from '@/app/dashboard/users/delete-form'
+import { convertParseISODate } from '@/utils/convertParseISODate'
 import ViewButton from '@/app/ui/dashboard/view-button/view-button'
-import { convertParseISODate } from '@/app/utils/convertParseISODate'
 
 type UsersTableProps = {
   data: Array<User>
