+++ src/app/ui/dashboard/mobile-sidebar/mobile-sidebar-client.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image'
 import { MdClose } from 'react-icons/md'
 
 import { TypesColors } from '@/enum/typeColors'
-import { menuItemsData } from '@/app/utils/menuItemsData'
+import { menuItemsData } from '@/utils/menuItemsData'
 import MenuLink from '@/app/ui/dashboard/menuLink/menulink'
 import { useSidebarMobileStore } from '@/store/sidebarMobileStore'
 import LogoutButton from '@/app/ui/dashboard/logout-button/logout-button'
