+++ src/app/ui/dashboard/sidebar/sidebar.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image'
 
 import { auth } from '@/app/auth'
 import { TypesColors } from '@/enum/typeColors'
-import { menuItemsData } from '@/app/utils/menuItemsData'
+import { menuItemsData } from '@/utils/menuItemsData'
 import MenuLink from '@/app/ui/dashboard/menuLink/menulink'
 import LogoutButton from '@/app/ui/dashboard/logout-button/logout-button'
