+++ src/utils/__tests__/convertStringToBoolean.spec.ts
