+++ src/utils/convertStringToBoolean.ts
