+++ src/utils/getItemsCache.ts
@@ -2,7 +2,7 @@ import { cache } from 'react'
 
 import { getTodos } from '@/app/services/api-todos'
 import { getUsers } from '@/app/services/api-users'
-import { getProducts } from '../services/api-products'
+import { getProducts } from '@/app/services/api-products'
 
 export const getTodosCache = cache(async () => {
   const todos = await getTodos()
